Fix issue with deleting user tasks from deleted workflow

+++ portal/node-server/job-executors/userTasks.ts
@@ -32,15 +32,16 @@ export async function modify(job: Job<BullMQ.UserTasks.Modify>): Promise<unknown
   let deletedCount = 0;
 
   // Clear PreExecuteEntries
-  await DatabaseWrites.productTransitions.deleteMany({
-    where: {
-      WorkflowUserId: null,
-      UserId: null,
-      ProductId: { in: productIds },
-      DateTransition: null
-    }
-  });
-
+  if (!project.DateArchived) {
+    await DatabaseWrites.productTransitions.deleteMany({
+      where: {
+        WorkflowUserId: null,
+        UserId: null,
+        ProductId: { in: productIds },
+        DateTransition: null
+      }
+    });
+  }
   job.updateProgress(20);
 
   let mapping: {
@@ -120,10 +121,10 @@ export async function modify(job: Job<BullMQ.UserTasks.Modify>): Promise<unknown
       job.updateProgress(job.data.operation.type === BullMQ.UserTasks.OpType.Delete ? 90 : 40);
     }
     for (let i = 0; i < products.length; i++) {
-      const product = products[i];
-      const snap = await Workflow.getSnapshot(product.Id);
       // Create tasks for all users that could perform this activity
       if (!project.DateArchived && job.data.operation.type !== BullMQ.UserTasks.OpType.Delete) {
+        const product = products[i];
+        const snap = await Workflow.getSnapshot(product.Id);
         const roleSet = new Set(
           (
             Workflow.availableTransitionsFromName(snap.state, snap.config)
@@ -153,11 +154,11 @@ export async function modify(job: Job<BullMQ.UserTasks.Modify>): Promise<unknown
         await DatabaseWrites.userTasks.createMany({
           data: createdTasks
         });
+        job.updateProgress(40 + ((i + 0.67) * 40) / products.length);
+        await DatabaseWrites.productTransitions.createMany({
+          data: await Workflow.transitionEntriesFromState(snap.state, products[i].Id, snap.config)
+        });
       }
-      job.updateProgress(40 + ((i + 0.67) * 40) / products.length);
-      await DatabaseWrites.productTransitions.createManyAndReturn({
-        data: await Workflow.transitionEntriesFromState(snap.state, products[i].Id, snap.config)
-      });
       job.updateProgress(40 + ((i + 1) * 40) / products.length);
     }
     job.updateProgress(80);
